Setting up bms to psql
