+++ back_end/App.py
@@ -1,5 +1,5 @@
 import threading
-from mqtt_db_subscriber import start_mqtt_subscriber
+from mqtt_subscriber import start_mqtt_subscriber
 from websocket import start_webserver
