+++ raspberry-pi/simulate.py
@@ -52,7 +52,7 @@ def publish(client):
         Args:
             client (mqtt_client): The publisher object connected to the AWS broker.
     """
-    with open("data/can_str_test.txt", "r") as file:
+    with open("data/bms_can_strings.txt", "r") as file:
         while True:
             msg = ""
             for line in file:
